+++ apps/dolly-frontend/src/main/js/src/pages/testnorgePage/search/SearchOptions.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
 import _get from 'lodash/get'
 import { FormikProps } from 'formik'
-import { OptionsSection } from './options/OptionsSection'
-import { Alder, AlderPaths } from './options/Alder'
-import { Identifikasjon, IdentifikasjonPaths } from './options/Identifikasjon'
-import { IdentSearch, IdentSearchPaths } from '~/pages/testnorgePage/search/options/IdentSearch'
-import { Adresser, AdresserPaths } from '~/pages/testnorgePage/search/options/Adresser'
-import { Nasjonalitet, NasjonalitetPaths } from '~/pages/testnorgePage/search/options/Nasjonalitet'
-import { Relasjoner, RelasjonerPaths } from '~/pages/testnorgePage/search/options/Relasjoner'
-import { Personstatus, PersonstatusPaths } from '~/pages/testnorgePage/search/options/Personstatus'
+import { Alder, AlderPaths } from './partials/Alder'
+import { Identifikasjon, IdentifikasjonPaths } from './partials/Identifikasjon'
+import { IdentSearch, IdentSearchPaths } from '~/pages/testnorgePage/search/partials/IdentSearch'
+import { Adresser, AdresserPaths } from '~/pages/testnorgePage/search/partials/Adresser'
+import { Nasjonalitet, NasjonalitetPaths } from '~/pages/testnorgePage/search/partials/Nasjonalitet'
+import { Relasjoner, RelasjonerPaths } from '~/pages/testnorgePage/search/partials/Relasjoner'
+import { Personstatus, PersonstatusPaths } from '~/pages/testnorgePage/search/partials/Personstatus'
+import { OptionsPanel } from './optionsPanel/OptionsPanel'
 
 export type SearchOptionsProps = {
 	formikBag: FormikProps<{}>
@@ -40,48 +40,55 @@ export const SearchOptions: React.FC<SearchOptionsProps> = (props: SearchOptions
 	return (
 		<>
 			<h2>Personinformasjon</h2>
-			<OptionsSection
+			<OptionsPanel
 				heading={'Fødsels- eller D-nummer'}
-				options={<IdentSearch formikBag={props.formikBag} />}
 				startOpen={true}
 				numSelected={getCount(IdentSearchPaths, props.formikBag)}
-			/>
-			<OptionsSection
+			>
+				<IdentSearch formikBag={props.formikBag} />
+			</OptionsPanel>
+			<OptionsPanel
 				heading={'Identifikasjon'}
-				options={<Identifikasjon formikBag={props.formikBag} />}
 				numSelected={getCount(IdentifikasjonPaths, props.formikBag)}
 				selectionColor={getSelectionColor(props.formikBag)}
-			/>
-			<OptionsSection
+			>
+				<Identifikasjon formikBag={props.formikBag} />
+			</OptionsPanel>
+			<OptionsPanel
 				heading={'Personstatus'}
-				options={<Personstatus />}
 				numSelected={getCount(PersonstatusPaths, props.formikBag)}
 				selectionColor={getSelectionColor(props.formikBag)}
-			/>
-			<OptionsSection
+			>
+				<Personstatus />
+			</OptionsPanel>
+			<OptionsPanel
 				heading={'Alder'}
-				options={<Alder />}
 				numSelected={getCount(AlderPaths, props.formikBag)}
 				selectionColor={getSelectionColor(props.formikBag)}
-			/>
-			<OptionsSection
+			>
+				<Alder />
+			</OptionsPanel>
+			<OptionsPanel
 				heading={'Adresser'}
-				options={<Adresser formikBag={props.formikBag} />}
 				numSelected={getCount(AdresserPaths, props.formikBag)}
 				selectionColor={getSelectionColor(props.formikBag)}
-			/>
-			<OptionsSection
+			>
+				<Adresser formikBag={props.formikBag} />
+			</OptionsPanel>
+			<OptionsPanel
 				heading={'Nasjonalitet'}
-				options={<Nasjonalitet />}
 				numSelected={getCount(NasjonalitetPaths, props.formikBag)}
 				selectionColor={getSelectionColor(props.formikBag)}
-			/>
-			<OptionsSection
+			>
+				<Nasjonalitet />
+			</OptionsPanel>
+			<OptionsPanel
 				heading={'Relasjoner'}
-				options={<Relasjoner formikBag={props.formikBag} />}
 				numSelected={getCount(RelasjonerPaths, props.formikBag)}
 				selectionColor={getSelectionColor(props.formikBag)}
-			/>
+			>
+				<Relasjoner formikBag={props.formikBag} />
+			</OptionsPanel>
 		</>
 	)
 }
