+++ apps/dolly-frontend/src/main/js/src/pages/testnorgePage/search/optionsPanel/OptionsPanel.less
@@ -98,17 +98,3 @@
 	border-bottom: 1px solid @color-bg-grey-border;
 	margin: 10px 0 15px 0;
 }
-
-.radio-group__options{
-	@import '@navikt/ds-css';
-}
-.navds-radio{
-	margin: 5px 0 10px 0;
-	&__input[type="radio"]{
-		height: 1em;
-		opacity: 100;
-	}
-	&__label{
-		margin-left: 5px;
-	}
-}
