+++ apps/dolly-frontend/src/main/js/src/pages/testnorgePage/search/optionsPanel/OptionsPanel.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import { useToggle } from 'react-use'
-import cn from 'classnames'
 import ExpandButton from '~/components/ui/button/ExpandButton'
 import './OptionsPanel.less'
 
@@ -18,25 +17,20 @@ export const OptionsPanel: React.FC<OptionsPanelProps> = ({
 	heading,
 	content,
 	children,
-	selectionColor,
-	numSelected,
+	selectionColor = 'blue',
+	numSelected = 0,
 }: OptionsPanelProps) => {
 	const [isOpen, toggleOpen] = useToggle(startOpen)
-
-	const panelClass = cn('options-panel', {
-		'options-panel-open': isOpen,
-	})
-
 	const renderContent = children ? children : content
 	const circleClassName = selectionColor
 		? 'options-panel-heading_' + selectionColor + '-circle'
 		: 'circle'
 
 	return (
-		<div className={panelClass}>
+		<div className="options-panel">
 			<div className="options-panel-heading" onClick={toggleOpen}>
 				<h2>{heading}</h2>
-				{selectionColor && numSelected && (
+				{selectionColor && numSelected > 0 && (
 					<div className={circleClassName}>
 						<p className="circle-text">{numSelected}</p>
 					</div>
