+++ apps/dolly-frontend/src/main/js/src/pages/testnorgePage/search/partials/Adresser.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import _get from 'lodash/get'
 import { FormikSelect } from '~/components/ui/form/inputs/select/Select'
 import { AdresseKodeverk } from '~/config/kodeverk'
-import { RadioGroupOptions } from '~/pages/testnorgePage/search/options/RadioGroupOptions'
+import { RadioGroupOptions } from '~/pages/testnorgePage/search/radioGroupOptions/RadioGroupOptions'
 import { FormikProps } from 'formik'
 import { yesNoOptions } from '~/pages/testnorgePage/utils'
