+++ apps/dolly-frontend/src/main/js/src/pages/testnorgePage/search/partials/Identifikasjon.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { FormikProps } from 'formik'
 import { FormikCheckbox } from '~/components/ui/form/inputs/checbox/Checkbox'
-import { RadioGroupOptions } from '~/pages/testnorgePage/search/options/RadioGroupOptions'
+import { RadioGroupOptions } from '~/pages/testnorgePage/search/radioGroupOptions/RadioGroupOptions'
 import { FormikSelect } from '~/components/ui/form/inputs/select/Select'
 
 interface IdentifikasjonProps {
