+++ apps/dolly-frontend/src/main/js/src/pages/testnorgePage/search/radioGroupOptions/RadioGroupOptions.less
@@ -1,0 +1,15 @@
+@import (reference) '~lessVars';
+
+.radio-group__options{
+  @import '@navikt/ds-css';
+}
+.navds-radio{
+  margin: 5px 0 10px 0;
+  &__input[type="radio"]{
+    height: 1em;
+    opacity: 100;
+  }
+  &__label{
+    margin-left: 5px;
+  }
+}
